feat: change kafka trigger to use consumption rates
Simply targetting a certain lag value leads to overscaling in some instances, and underscaling in others

For example, if we drop under the lag threshold we might scale down, but the consumers may not be able to keep up with the new number of replicas, so we'd build lag then scale up again later

This change makes the KafkaLagTriggerProcessor target throughput on the topic instead of an absolute lag value. If we're under the lag threshold we use historic throughput values to calculate whether we're under or over the throughput into the head of the topic itself.  If we're over the lag threshold we use an 'sla' target to calculate the correct number of replicas needed to clear own the lag in a given period of time (default 10 minutes)